Design Kotlin Item Card and fixed NetworkModule.kt

+++ app/src/main/java/com/example/myandroidbook/domain/model/KotlinModel.kt
@@ -14,7 +14,7 @@ data class KotlinModel (
     val image: String,
     val about: String,
     val tags : List<String>,
-    val ranking: Int,
+    val ranking: Double,
     val yearRelease: Int,
     val difficulty : String
 )
